Simplify and reformat AddSystemForm
It does not need to be a form.

Also moves components around to start to match updated design
