--- clients/admin-ui/src/features/system/constants.ts
@@ -1,2 +1,0 @@
-export const ADD_SYSTEM_DESCRIPTION =
-  "Systems describe any services that store or process data for your organization, including third-party APIs, web applications, databases, and data warehouses. System discovery allows Fides to identify and build a data map from the list of systems that exist within your organization.";
